refactor: use orjson
