+++ docarray/base_doc/base_node.py
@@ -1,8 +1,10 @@
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, TypeVar, Optional, Type
 
 if TYPE_CHECKING:
     from docarray.proto import NodeProto
+
+T = TypeVar('T')
 
 
 class BaseNode(ABC):
@@ -10,6 +12,8 @@ class BaseNode(ABC):
     A DocumentNode is an object than can be nested inside a Document.
     A Document itself is a DocumentNode as well as prebuilt type
     """
+
+    _proto_type_name: Optional[str] = None
 
     @abstractmethod
     def _to_node_protobuf(self) -> 'NodeProto':
@@ -20,3 +24,15 @@ class BaseNode(ABC):
         :return: the nested item protobuf message
         """
         ...
+
+    @classmethod
+    @abstractmethod
+    def from_protobuf(cls: Type[T], pb_msg: T) -> T:
+        ...
+
+    def _docarray_to_json_compatible(self):
+        """
+        Convert itself into a json compatible object
+        :return: a representation of the tensor compatible with orjson
+        """
+        return self
